+++ app/src/main/java/nextstep/jwp/controller/AbstractController.java
@@ -1,0 +1,17 @@
+package nextstep.jwp.controller;
+
+import nextstep.jwp.http.request.HttpRequest;
+import nextstep.jwp.http.response.HttpResponse;
+
+public abstract class AbstractController implements Controller {
+
+    @Override
+    public void service(HttpRequest request, HttpResponse response) throws Exception {
+        // http method 분기문
+    }
+
+    protected void doPost(HttpRequest request, HttpResponse response) throws Exception { /* NOOP */ }
+
+    protected void doGet(HttpRequest request, HttpResponse response) throws Exception { /* NOOP */ }
+}
+
