+++ app/src/main/java/nextstep/jwp/controller/Controller.java
@@ -1,0 +1,8 @@
+package nextstep.jwp.controller;
+
+import nextstep.jwp.http.request.HttpRequest;
+import nextstep.jwp.http.response.HttpResponse;
+
+public interface Controller {
+    void service(HttpRequest request, HttpResponse response) throws Exception;
+}
