+++ app/src/main/java/nextstep/jwp/http/request/RequestHeader.java
@@ -1,4 +1,4 @@
-package nextstep.jwp;
+package nextstep.jwp.http.request;
 
 import java.io.BufferedReader;
 import java.io.IOException;
