+++ app/src/main/java/nextstep/jwp/http/response/HttpResponse.java
@@ -1,11 +1,11 @@
-package nextstep.jwp;
+package nextstep.jwp.http.response;
 
-public class Response {
+public class HttpResponse {
     private final ResponseLine responseLine;
     private final ResponseHeader responseHeader;
     private final ResponseBody responseBody;
 
-    public Response(ResponseLine responseLine, ResponseHeader responseHeader, ResponseBody responseBody) {
+    public HttpResponse(ResponseLine responseLine, ResponseHeader responseHeader, ResponseBody responseBody) {
         this.responseLine = responseLine;
         this.responseHeader = responseHeader;
         this.responseBody = responseBody;
