+++ app/src/main/java/nextstep/jwp/http/response/ResponseLine.java
@@ -1,4 +1,4 @@
-package nextstep.jwp;
+package nextstep.jwp.http.response;
 
 import java.util.Arrays;
